DevFreela.Application/Queries/GetAllUsers/GetAllUsersQueryHandler.cs
using DevFreela.Core.Entities;
using DevFreela.Core.Repositories;
using MediatR;

namespace DevFreela.Application.Queries.GetAllUsers;

public class GetAllUsersQueryHandler : IRequestHandler<GetAllUsersQuery, List<User>>
{
    private readonly IUserRepository _userRepository;

    public GetAllUsersQueryHandler(IUserRepository userRepository)
    {
        _userRepository = userRepository;   
    }

    public async Task<List<User>> Handle(GetAllUsersQuery request, CancellationToken cancellationToken) => await _userRepository.GetAllAsync();

}
